hmc.py: add profile option, timeouts, more traces
Added a function to get the lpar current profile to be booted.

During Linux boot if crash/panic or any similar traces are detected the job
exits. Due to this logic sometimes complete trace is not captured. Increase
the timeout to 5 minutes so that entire trace can be captured in the console log

Due to slow network smsmenu is not listed soon when mkvterm command is run,
so increased more wait time to get to sms menu

fix sms main page timing issue, fix is to see if the page recieved invalid
key due to timing issue and than preset the page and continue sending key

More traces added to parser, now we have new kernel failures that will
be parsed

Signed-off-by: Abdul Haleem <abdhalee@linux.vnet.ibm.com>